Add quantum hybrid network into separate files and add mnist dataset loader
