+++ pyproject.toml
@@ -20,6 +20,8 @@ typer = "^0.9.0"
 rich = "^13.7.0"
 
 [tool.poetry.scripts]
+ipsem2025-train = "ipsem2025_license_plate.qnet.cli:main"
+ipsem2025 = "ipsem2025_license_plate.cli:main"
 ipsem2025-dataset = "ipsem2025_license_plate.datasets.cli:main"
 
 [tool.poetry.group.format.dependencies]
