+++ src/ipsem2025_license_plate/qnet/__init__.py
@@ -1,0 +1,16 @@
+"""Quantum-classical hybrid neural network package."""
+
+from .model import HybridModel
+from .train import train_model, train_hybrid_model
+from .test import evaluate_model
+from .utils import save_model, load_model, get_hardware_info
+
+__all__ = [
+    'HybridModel',
+    'train_model',
+    'train_hybrid_model',
+    'evaluate_model',
+    'save_model',
+    'load_model',
+    'get_hardware_info'
+]
