--- tests/test_ipsem2025_license_plate.py
@@ -1,6 +1,0 @@
-"""Basic tests on the package."""
-
-
-def test_dummy():
-    """A dummy positive test."""
-    assert True
